requirements.txt
streamlit
ebooklib
streamlit-sortables
Pillow